requirements.txt
streamlit
requests
python-dotenv
pandas
folium
streamlit-folium
pyperclip
streamlit-js-eval